requirements.txt
streamlit==1.30.0
pandas==2.1.0
python-docx==1.0.0